Create OpenWatchFilePandas.py
This code can open an export file. I set it to skip the first 153 lines, which is hopefully the same number in each subject export file. When it prints the csv, the columns are sometimes put underneath, probably because there is a limit on the screen size. I don't know what impact this will have in the future when we try to join csv files together.