Create New Watch file
This code kinda of works. It is able to read the file names and choose the files based on subject #, but for some reason, access to the raw data files is denied. (PermissionError: [Errno 13] Permission denied: 'C:\\Users\\Isabelle\\Python\\Individual_Watch_Exports')

The problem is probably in line 22. When I take out "path", it says that the csv file can't be found. That is probably because it is searching for the csvs in C:\Users\Isabelle\Python\, where the code is saved. But, I can't move the code to the "Individual_Watch_Exports" folder because if I do, the part that identifies the subject_id will not work because there will be a file that doesn't follow the pattern of Subject_#_etc.csv