Update WeekendWatchLabel.py
It works!
It labels whether or not the time stamp is between 5pm on Friday to 5pm on Sunday. 1=yes, 0= no. It is built on the concatenating code, but should not take all that long since it uses a dict to open the files

(Edit: In changing Saturday to Sunday, I accidentally skipped Saturday. Saturday has been added back in to the weekend)